+++ real_shit/libft/functions/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hassanAsarhan@outlook.com>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/09 18:09:25 by hsarhan           #+#    #+#             */
-/*   Updated: 2022/05/14 10:42:35 by hsarhan          ###   ########.fr       */
+/*   Updated: 2022/05/15 13:12:30 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 {
 	char		*dst_bytes;
 	const char	*src_bytes;
-	int			i;
+	size_t		i;
 
 	if (len == 0)
 		return (dst);
