+++ real_shit/libft/functions/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hassanAsarhan@outlook.com>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/12 14:07:37 by hsarhan           #+#    #+#             */
-/*   Updated: 2022/05/13 19:57:11 by hsarhan          ###   ########.fr       */
+/*   Updated: 2022/05/15 12:55:53 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	len;
-	char			*str;
-	unsigned int	i;
+	size_t	len;
+	char	*str;
+	size_t	i;
 
 	len = ft_strlen(s);
 	str = malloc(sizeof(char) * (len + 1));
